rename and lint
